Refactor stock system with confirmation for deletions

+++ main.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from functions import cadastrar_produto, excluir_produto, entrada_saida_estoque, pedir_valor_unico, pedir_tipo_variavel
+from functions import cadastrar_produto, excluir_produto, entrada_estoque, saida_estoque, pedir_valor_unico, pedir_tipo_variavel, confirmar_id_unico
 
 df_produtos = pd.DataFrame({
     'ID': ['NJQ1627','NMT4639','MPV3588','CBN2845'],
@@ -11,7 +11,8 @@ df_produtos = pd.DataFrame({
 df_produtos = df_produtos.set_index('ID')
 
 while True:
-    
+
+    print('\n== SISTEMA DE ESTOQUE ==')
     print('\nMenu Inicial:')
     print('1 - Cadastrar Produto')
     print('2 - Excluir Produto')
@@ -21,11 +22,8 @@ while True:
     print('0 - Sair\n')
     
     opcao = input('Selecione a sua opção: ')
-
-    if opcao == '0':
-        break
     
-    elif opcao == '1':
+    if opcao == '1':
         
         nome = pedir_valor_unico(df_produtos['PRODUTO'], 'Produto: ')
         preco = pedir_tipo_variavel(float, 'Valor: ')
@@ -36,24 +34,33 @@ while True:
     elif opcao == '2':
 
         c_id = input('Código ID: ')
-        df_produtos = excluir_produto(c_id, df_produtos)
-        
+        if confirmar_id_unico(c_id, df_produtos) == True:
+
+            conf = input(f"Deseja mesmo deletar produto {c_id}? \nDigite: Sim / Não\nResposta: ").lower()
+            excluir_produto(conf, c_id, df_produtos)
+            df_produtos = excluir_produto(c_id, df_produtos)
+
     elif opcao == '3':
 
         c_id = input('Código ID: ')
-        quant = pedir_tipo_variavel(int, 'Informe o número de entrada de estoque: ')
+        if confirmar_id_unico(c_id, df_produtos) == True:
 
-        df_produtos = entrada_saida_estoque('entrada', c_id, quant, df_produtos)
+            quant = pedir_tipo_variavel(int, 'Informe o número de entrada de estoque: ')
+            df_produtos = entrada_estoque('entrada', c_id, quant, df_produtos)
         
     elif opcao == '4':
 
         c_id = input('Código ID: ')
-        quant = pedir_tipo_variavel(int, 'Informe o número de saída de estoque: ')
-        
-        df_produtos = entrada_saida_estoque('saida', c_id, quant, df_produtos)
+        if confirmar_id_unico(c_id, df_produtos) == True:
+
+            quant = pedir_tipo_variavel(int, 'Informe o número de saída de estoque: ')
+            df_produtos = saida_estoque('saida', c_id, quant, df_produtos)
         
     elif opcao == '5':
         print(f'\n{df_produtos}\n')
+
+    elif opcao == '0':
+        break
         
     else:
         print('Entrada inválida, utilize números de 0 a 5!')
